Process cn readme also

+++ rake_lib/files.rb
@@ -23,11 +23,13 @@ module AppiumIo
       raise "Source must exist. #{source}" unless exist?(source)
 
       dest = expand_path dest
+      raise 'source must not equal dest' if source == dest
       # copy_entry '/a/b/c.png', '/e/f' => copies to /e/f/c.png
       dest = join(dest, basename(source)) if !directory?(source) && directory?(dest)
 
       dir_to_make = directory?(dest) ? dest : dirname(dest)
       mkdir_p dir_to_make
+      raise "Dest must exist. #{dir_to_make}" unless exist?(dir_to_make)
 
       FileUtils.copy_entry source, dest, preserve = false,
                            dereference_root       = false,
